Pull Curry into MetaMap
