+++ lib/ops.js
@@ -5,21 +5,8 @@ var __extends = (this && this.__extends) || function (d, b) {
     d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 };
 var frames_1 = require("./frames");
-exports.Curry = function (func, source) {
-    return function (block) {
-        return func(source, block);
-    };
-};
 var iterators_1 = require("./ops/iterators");
 exports.MetaMap = iterators_1.MetaMap;
-var FrameCurry = (function (_super) {
-    __extends(FrameCurry, _super);
-    function FrameCurry(func) {
-        _super.call(this);
-    }
-    return FrameCurry;
-}(frames_1.Frame));
-exports.FrameCurry = FrameCurry;
 var FrameOps = (function (_super) {
     __extends(FrameOps, _super);
     function FrameOps(context) {
@@ -31,6 +18,4 @@ var FrameOps = (function (_super) {
     return FrameOps;
 }(frames_1.Frame));
 exports.FrameOps = FrameOps;
-exports.Ops = new FrameOps({
-    "&&": new FrameCurry(iterators_1.MetaMap),
-});
+exports.Ops = new FrameOps({});
