+++ src/ops.ts
@@ -1,22 +1,7 @@
 import { Context, Frame, FrameArray, FrameExpr } from "./frames";
 
-export interface ICurryFunction extends Function {
-  (source: Frame, block: Frame): Frame;
-}
-
-export const Curry = (func: ICurryFunction, source: Frame) => {
-  return (block: Frame) => {
-    return func(source, block);
-  };
-};
-
 import { MetaMap, MetaMapExpr } from "./ops/iterators";
 
-export class FrameCurry extends Frame {
-  constructor(func: ICurryFunction) {
-    super();
-  }
-}
 
 export class FrameOps extends Frame {
   constructor(context: Context) {
@@ -29,7 +14,6 @@ export class FrameOps extends Frame {
 }
 
 export const Ops = new FrameOps({
-  "&&": new FrameCurry(MetaMap),
 });
 
 export { MetaMap };
